Refactor HomePage structure and rename files for consistency
